feat: seller page create and read function developed

+++ src/pages/NewProduct.tsx
@@ -1,6 +1,6 @@
 import { useState, ChangeEvent } from "react";
 import FileUpload, { uploadFiles } from "../components/FileUpload";
-import { addDoc, collection } from "firebase/firestore";
+import { addDoc, collection, Timestamp } from "firebase/firestore";
 import { db, storage } from "../api/firebase";
 import { RiDeleteBin5Fill } from "react-icons/ri";
 import { deleteObject, ref } from "firebase/storage";
@@ -13,7 +13,7 @@ interface Product {
   description: string;
   options: string[];
   imageUrls?: string[];
-  // 다른 필요한 속성들...
+  createdAt: Timestamp;
 }
 
 export default function NewProduct() {
@@ -24,6 +24,7 @@ export default function NewProduct() {
     category: "",
     description: "",
     options: [],
+    createdAt: Timestamp.now(),
   });
   const [selectedFiles, setSelectedFiles] = useState<File[]>([]);
   const [uploadedUrls, setUploadedUrls] = useState<string[]>([]);
@@ -56,9 +57,12 @@ export default function NewProduct() {
 
   // Firebase Storage에서 이미지를 삭제하는 함수
   const deleteImage = async (imageUrl: string) => {
-    const imageRef = ref(storage, imageUrl);
-
     try {
+      console.log("삭제할 이미지 URL:", imageUrl);
+      if (!imageUrl) {
+        throw new Error("올바르지 않은 이미지 URL입니다.");
+      }
+      const imageRef = ref(storage, imageUrl);
       await deleteObject(imageRef);
       console.log("이미지가 성공적으로 삭제되었습니다.");
     } catch (error) {
@@ -67,11 +71,17 @@ export default function NewProduct() {
   };
 
   // 이미지 삭제 핸들러
-  const handleRemoveImage = (index: number) => {
+  const handleRemoveImage = async (index: number) => {
     const imageUrlToRemove = uploadedUrls[index];
 
+    // URL이 설정되었는지 확인
+    if (!imageUrlToRemove) {
+      console.error("이미지 URL이 설정되지 않았습니다.");
+      return;
+    }
+
     // Firebase Storage에서 이미지 삭제
-    deleteImage(imageUrlToRemove);
+    await deleteImage(imageUrlToRemove);
 
     // Client 단 이미지 업로드 상태 업데이트
     setSelectedFiles((prevFiles) => prevFiles.filter((_, i) => i !== index));
@@ -91,7 +101,11 @@ export default function NewProduct() {
     setIsUploading(true);
 
     try {
-      const productData = { ...product, imageUrls: uploadedUrls };
+      const productData = {
+        ...product,
+        imageUrls: uploadedUrls,
+        createdAt: Timestamp.now().toDate(),
+      };
 
       // Firestore에 제품 데이터 저장
       await addDoc(collection(db, "products"), productData);
@@ -104,6 +118,7 @@ export default function NewProduct() {
         category: "",
         description: "",
         options: [],
+        createdAt: Timestamp.now(),
       });
       setSelectedFiles([]);
       setUploadedUrls([]);
